move CIContext to bases module to avoid circular deps
