+++ tools/ci_connector_ops/ci_connector_ops/pipelines/bases.py
@@ -28,6 +28,15 @@ if TYPE_CHECKING:
     from ci_connector_ops.pipelines.contexts import ConnectorContext, PipelineContext
 
 
+class CIContext(str, Enum):
+    """An enum for Ci context values which can be ["manual", "pull_request", "nightly_builds"]."""
+
+    MANUAL = "manual"
+    PULL_REQUEST = "pull_request"
+    NIGHTLY_BUILDS = "nightly_builds"
+    MASTER = "master"
+
+
 class StepStatus(Enum):
     """An Enum to characterize the success, failure or skipping of a Step."""
