+++ tools/ci_connector_ops/ci_connector_ops/pipelines/commands/airbyte_ci.py
@@ -5,7 +5,7 @@
 """This module is the CLI entrypoint to the airbyte-ci commands."""
 
 import click
-from ci_connector_ops.pipelines.contexts import CIContext
+from ci_connector_ops.pipelines.bases import CIContext
 from ci_connector_ops.pipelines.utils import (
     get_current_epoch_time,
     get_current_git_branch,
